quite a lot things again
tagManager prevented multiple news with the same tags and tile present
at the same place!

generalized search condition composer!
not it can prepare condition and criteria property of $this for
tagsManagerCheckCondition.sql

new way of writing statements on the fly
tagsManagerCheckStatement.'fbid'.sql

sql code composed on the fly not synced